Add mockRef test util, separate out TextInput/TextInputField tests
