+++ src/select/__tests__/Select.test.js
@@ -1,14 +1,13 @@
 import React from 'react'
 import { render } from '@testing-library/react'
 import { Select } from '../'
+import { mockRef } from '../../test/utils'
 
 const makeSelectFixture = (props = {}) => <Select data-testid="select" {...props} />
 
 describe('Select', () => {
   it('should forward ref to underlying <select />', () => {
-    const ref = {
-      current: jest.fn()
-    }
+    const ref = mockRef()
 
     render(makeSelectFixture({ ref }))
